% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fit_measures}
\alias{fit_measures}
\title{Get different measures of SEM fit}
\usage{
fit_measures(model_obj, ...)
}
\arguments{
\item{model_obj}{Fitted \code{thurstem} model object.}

\item{...}{Extra arguments to be pass on to \code{\link[lavaan]{fitMeasures}} from
\pkg{lavaan}.}
}
\value{
A \code{lavaan}-esque summary of the various SEM fit measures.
See \code{\link[lavaan]{fitMeasures}}.
}
\description{
Get different measures of SEM fit
}
