% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rank_mod.R
\name{fit_rank_mod}
\alias{fit_rank_mod}
\title{Fit a Thurstone Rank Model}
\usage{
fit_rank_mod(data, pairs_cols = NULL, form = NULL, mclass = "V")
}
\arguments{
\item{data}{A data frame containing the pairs data and any covariates.}

\item{pairs_cols}{A vector of column indices or names specifying the pairwise comparison columns in the data.
If \code{NULL}, all columns are used.}

\item{form}{An optional formula specifying the covariates to include in the model. If \code{NULL}, then the
\emph{intercept only} model will be fitted.}

\item{mclass}{A character string specifying the Thurstone model class to fit. Currently, only "V" is implemented.}
}
\value{
A list of class \code{"thurstem"} containing:
\item{mfit}{The fitted lavaan model object.}
\item{df_pvals}{A data frame of chi-square test statistics, degrees of freedom, and p-values for each covariate.}
\item{mclass}{The specified Thurstone model class.}
}
\description{
This function fits a Thurstone rank model to a dataset.
The function allows for optional inclusion of covariates and handles the conversion of pairwise
ranking data into a form suitable for fitting using the \code{lavaan} package.
}
