% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.thurstem}
\alias{summary.thurstem}
\title{Generic summary for a \code{thurstem} object}
\usage{
\method{summary}{thurstem}(model_obj)
}
\arguments{
\item{model_obj}{A fitted \code{thurstem} object.}
}
\value{
\code{lavaan}-esque summary, including sections for
\itemize{
\item A summary of model fitting procedures and model fit
\item A summary of the latent variables and the factor loadings
\item A summary of the regression component of the model, modeling
the intercepts of the latent variables
\item A summary of the threshold parameters, for which \eqn{1 - \Phi(\tau_\ell)},
where \eqn{\Phi()} is the standard normal CDF, gives the estimated probability
of comparison \eqn{\ell}, that is, item \eqn{i} ranked ahead of item \eqn{j}.
\item A summary of the variances and covariances of the latent variables. Note that
these will be fixed in most cases.
\item Scales of \eqn{y^*}. Note that the user will not often find use for this section.
\item A summary of the defined parameters. This is the section in which users will find
information on marginal means of the latent utility distributions.
}
}
\description{
Generic summary for a \code{thurstem} object
}
