% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.thurstem}
\alias{summary.thurstem}
\title{Generic summary for \code{thurstem} object}
\usage{
\method{summary}{thurstem}(model_obj)
}
\arguments{
\item{model_obj}{Fitted \code{thurstem} object}
}
\value{
\code{lavaan}-esque summary
}
\description{
Generic summary for \code{thurstem} object
}
